requirements.txt
langgraph>=0.2.28
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-aws>=0.2.0
langchain-ollama>=0.2.0
langchain-huggingface>=0.1.0
langchain-community>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
langsmith>=0.1.0

chromadb
sentence-transformers
pypdf
python-docx
beautifulsoup4
requests